+++ src/SoftTissue.DataContract/ExperimentalAnalysis/AnalyzeAndExtendResults/AnalyzeAndExtendResultsRequest.cs
@@ -1,6 +1,6 @@
 ﻿using SoftTissue.DataContract.OperationBase;
 
-namespace SoftTissue.DataContract.Experimental.AnalyzeAndExtendResults
+namespace SoftTissue.DataContract.ExperimentalAnalysis.AnalyzeAndExtendResults
 {
     /// <summary>
     /// It represents the request content to AnalyzeAndPredictResults operation.
@@ -16,11 +16,6 @@ namespace SoftTissue.DataContract.Experimental.AnalyzeAndExtendResults
         /// The file URI.
         /// </summary>
         public string FileUri { get; set; }
-
-        /// <summary>
-        /// True, if should use the time step from experimental file. False, otherwise.
-        /// </summary>
-        public bool UseTimeStepFromFile { get; set; }
 
         /// <summary>
         /// Time step.